+++ backEnd/src/main/java/com/wellnessapp/demo/CreatorRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
 
-public interface CreatorRepository extends MongoRepository<Creator, Long> {
+public interface CreatorRepository extends MongoRepository<Creator, Integer> {
     // possible functions for querrying the db
     public Creator findByUsername(String username);
     public List<Creator> findByLevel(int level);
